feat: click timeline labels to view that section
