+++ eslint.config.mjs
@@ -40,7 +40,6 @@ const eslintConfig = defineConfig([
       "react/hook-use-state": "warn",
       "react/jsx-boolean-value": "warn",
       "react/jsx-handler-names": "warn",
-      "react/jsx-no-bind": "error",
       "react/jsx-pascal-case": "error",
       "react/jsx-sort-props": "warn",
       "react/jsx-uses-react": "error",
